requirements.txt
pandas
scikit-learn
joblib
bs4
utils
requests
lxml
sqlalchemy
psycopg2-binary 
python-dotenv 
openpyxl